Honey/ExpenseTrackerMobile/requirements.txt
kivy==2.2.1
mysql-connector-python==8.2.0
Pillow==10.1.0
werkzeug==3.0.1
